LayoutTests/storage/indexeddb/setVersion-undefined-expected.txt

Test IndexedDB version string is required

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


indexedDB = self.indexedDB || self.webkitIndexedDB || self.mozIndexedDB || self.msIndexedDB || self.OIndexedDB;

indexedDB.open(name)
db = event.target.result
PASS db.setVersion(); threw exception TypeError: Not enough arguments.
PASS successfullyParsed is true

TEST COMPLETE
